POC: Logs won't be moved for time being
