+++ frigate/api/fastapi/new_app.py
@@ -2,8 +2,9 @@ import logging
 
 from fastapi import FastAPI
 
+from .. import app as main_app
 from .. import media, preview
-from .routers import logs, test
+from .routers import test
 
 logger = logging.getLogger(__name__)
 
@@ -29,9 +30,9 @@ def create_fastapi_app(frigate_config, detected_frames_processor):
     logger.info("Starting FastAPI app")
     app = FastAPI(debug=False, tags_metadata=tags_metadata)
     app.include_router(test.router)
-    app.include_router(logs.router)
     app.include_router(preview.router)
     app.include_router(media.router)
+    app.include_router(main_app.router)
     app.frigate_config = frigate_config
     app.detected_frames_processor = detected_frames_processor
     app.camera_error_image = None
